src/requirements.txt
pytest
mock
requests
mysql-connector
python-dateutil
coverage
boto3
cerberus
moto
botocore
requests-mock
pytest-env